+++ lib/ngtcp2_err.c
@@ -105,7 +105,7 @@ const char *ngtcp2_strerror(int liberr) {
 
 int ngtcp2_err_is_fatal(int liberr) { return liberr < NGTCP2_ERR_FATAL; }
 
-uint16_t ngtcp2_err_infer_quic_transport_error_code(int liberr) {
+uint64_t ngtcp2_err_infer_quic_transport_error_code(int liberr) {
   switch (liberr) {
   case 0:
     return NGTCP2_NO_ERROR;
